add configs for heart disease dataset
