--- examples/resources/config_gc/flamby/clients.yaml
@@ -1,74 +1,0 @@
-clients:
-  - endpoint_id: "17b4d21f-3f70-47a8-af4d-736c5e89e0d6"
-    train_configs:
-      # Device
-      device: "cpu"
-      # Logging and outputs
-      logging_id: "Client1"
-      logging_output_dirname: "./output"
-      logging_output_filename: "result"
-
-    # Local dataset
-    data_configs:
-      dataset_path: "./resources/dataset/flamby_dataset.py"
-      dataset_name: "get_flamby"
-      dataset_kwargs:
-        dataset: "HeartDisease"
-        num_clients: 4
-        client_id: 0
-
-  - endpoint_id: "270c774d-f081-4445-a6d2-94d6ec4f1f8f"
-    train_configs:
-      # Device
-      device: "cpu"
-      # Logging and outputs
-      logging_id: "Client2"
-      logging_output_dirname: "./output"
-      logging_output_filename: "result"
-
-    # Local dataset
-    data_configs:
-      dataset_path: "./resources/dataset/flamby_dataset.py"
-      dataset_name: "get_flamby"
-      dataset_kwargs:
-        dataset: "HeartDisease"
-        num_clients: 4
-        client_id: 1
-
-  - endpoint_id: "2004f578-b796-4d78-8aab-1b4873521bde"
-    train_configs:
-      # Device
-      device: "cpu"
-      # Logging and outputs
-      logging_id: "Client3"
-      logging_output_dirname: "./output"
-      logging_output_filename: "result"
-
-    # Local dataset
-    data_configs:
-      dataset_path: "./resources/dataset/flamby_dataset.py"
-      dataset_name: "get_flamby"
-      dataset_kwargs:
-        dataset: "HeartDisease"
-        num_clients: 4
-        client_id: 2
-
-  - endpoint_id: "db20a7f0-4b49-4ae5-98bc-34166df5fbcc"
-    train_configs:
-      # Device
-      device: "cpu"
-      # Logging and outputs
-      logging_id: "Client4"
-      logging_output_dirname: "./output"
-      logging_output_filename: "result"
-
-    # Local dataset
-    data_configs:
-      dataset_path: "./resources/dataset/flamby_dataset.py"
-      dataset_name: "get_flamby"
-      dataset_kwargs:
-        dataset: "HeartDisease"
-        num_clients: 4
-        client_id: 3
-
-
