+++ examples/resources/config_gc/flamby/heart_disease/clients.yaml
@@ -1,0 +1,38 @@
+clients:
+  - endpoint_id: "ed4a1881-120e-4f67-88d7-876cd280feef"
+    client_id: "Client1"
+    train_configs:
+      # Device
+      device: "cpu"
+      # Logging and outputs
+      logging_id: "Client1"
+      logging_output_dirname: "./output"
+      logging_output_filename: "result"
+
+    # Local dataset
+    data_configs:
+      dataset_path: "./resources/dataset/flamby_dataset.py"
+      dataset_name: "get_flamby"
+      dataset_kwargs:
+        dataset: "HeartDisease"
+        num_clients: 4
+        client_id: 0
+
+  - endpoint_id: "762629a0-f3b3-44b5-9acf-2f9b0ab9310f"
+    client_id: "Client2"
+    train_configs:
+      # Device
+      device: "cpu"
+      # Logging and outputs
+      logging_id: "Client2"
+      logging_output_dirname: "./output"
+      logging_output_filename: "result"
+
+    # Local dataset
+    data_configs:
+      dataset_path: "./resources/dataset/flamby_dataset.py"
+      dataset_name: "get_flamby"
+      dataset_kwargs:
+        dataset: "HeartDisease"
+        num_clients: 4
+        client_id: 1
