+++ examples/resources/config_gc/flamby/heart_disease/server_fedavg.yaml
@@ -4,49 +4,28 @@ client_configs:
     trainer: "VanillaTrainer"
     mode: "step"
     num_local_steps: 100
-    optim: "SGD"
+    optim: "Adam"
     optim_args:
-      lr: 0.01
+      lr: 0.001
     # Loss function
-    loss_fn_path: "./resources/loss/flamby_heart_disease.py"
+    loss_fn_path: "./resources/loss/bceloss.py"
     loss_fn_name: "BaselineLoss"
     # Client validation
     do_validation: True
     do_pre_validation: True
-    metric_path: "./resources/metric/flamby_heart_disease.py"
+    metric_path: "./resources/metric/binary_acc.py"
     metric_name: "metric"
     # Differential privacy
     use_dp: False
     epsilon: 1
-    clip_grad: True
-    clip_value: 0.5
+    clip_grad: False
+    clip_value: 1
     clip_norm: 1
     # Data loader
-    train_batch_size: 16
-    val_batch_size: 16
+    train_batch_size: 4
+    val_batch_size: 4
     train_data_shuffle: True
     val_data_shuffle: False
-    
-  model_configs:
-    model_path: "./resources/model/flamby_heart_disease.py"
-    model_name: "Baseline"
-    # model_kwargs:
-    #   input_size: 13
-    #   num_classes: 1
-  
-  # comm_configs will be copied to server_configs as well
-  comm_configs:
-    compressor_configs:
-      enable_compression: False
-      # Used if enable_compression is True
-      lossy_compressor:  "SZ2Compressor"
-      lossless_compressor: "blosc"
-      error_bounding_mode: "REL"
-      error_bound: 1e-3
-      param_cutoff: 1024
-    # globus_compute_configs:
-    #   s3_bucket: "appflx-models"
-    #   s3_creds_file: "./resources/config_gc/creds/credential.csv"
 
   data_readiness_configs:
     generate_dr_report: True
@@ -78,27 +57,31 @@ client_configs:
         feature_statistics_plot: True
       combine:
         feature_space_distribution: True
+    
+  model_configs:
+    model_path: "./resources/model/flamby_heart_disease.py"
+    model_name: "Baseline"
+  
+  comm_configs:
+    compressor_configs:
+      enable_compression: False
+      # Used if enable_compression is True
+      lossy_compressor:  "SZ2Compressor"
+      lossless_compressor: "blosc"
+      error_bounding_mode: "REL"
+      error_bound: 1e-3
+      param_cutoff: 1024
 
 server_configs:
-  scheduler: "CompassScheduler"
+  scheduler: "SyncScheduler"
   scheduler_kwargs:
-    num_clients: 4
+    num_clients: 2
     same_init_model: True
-    max_local_steps: 100
-    min_local_steps: 20
-    speed_momentum: 0.9
-    latest_time_factor: 1.5
-  aggregator: "FedCompassAggregator"
+  aggregator: "FedAvgAggregator"
   aggregator_kwargs:
-    weights: "equal"
-    staleness_fn: "polynomial"
-    staleness_fn_kwargs:
-      a: 0.5
-    alpha: 0.9
-    gradient_based: True
-    num_clients: 4
+    client_weights_mode: "equal"
   device: "cpu"
-  num_global_epochs: 17
+  num_global_epochs: 20
   logging_output_dirname: "./output"
   logging_output_filename: "result"
   comm_configs:
