+++ examples/resources/config_gc/flamby/heart_disease/server_fedcompass.yaml
@@ -8,49 +8,40 @@ client_configs:
     optim_args:
       lr: 0.001
     # Loss function
-    loss_fn_path: "./resources/loss/celoss.py"
-    loss_fn_name: "CELoss"
+    loss_fn_path: "./resources/loss/bceloss.py"
+    loss_fn_name: "BaselineLoss"
     # Client validation
     do_validation: True
     do_pre_validation: True
-    metric_path: "./resources/metric/acc.py"
-    metric_name: "accuracy"
+    metric_path: "./resources/metric/binary_acc.py"
+    metric_name: "metric"
+    send_gradient: True
     # Differential privacy
     use_dp: False
     epsilon: 1
     clip_grad: False
     clip_value: 1
     clip_norm: 1
-    # Data format
-    send_gradient: True
     # Data loader
-    train_batch_size: 64
-    val_batch_size: 64
+    train_batch_size: 4
+    val_batch_size: 4
     train_data_shuffle: True
     val_data_shuffle: False
-    
+
   model_configs:
-    model_path: "./resources/model/cnn.py"
-    model_name: "CNN"
-    model_kwargs:
-      num_channel: 1
-      num_classes: 10
-      num_pixel: 28
+    model_path: "./resources/model/flamby_heart_disease.py"
+    model_name: "Baseline"
   
-  # comm_configs will be copied to server_configs as well
   comm_configs:
     compressor_configs:
-      enable_compression: True
+      enable_compression: False
       # Used if enable_compression is True
       lossy_compressor:  "SZ2Compressor"
       lossless_compressor: "blosc"
       error_bounding_mode: "REL"
       error_bound: 1e-3
       param_cutoff: 1024
-    globus_compute_configs:
-      s3_bucket: "appflx-models"
-      s3_creds_file: "./resources/config_gc/creds/credential.csv"
-    
+
   data_readiness_configs:
     generate_dr_report: True
     output_dirname: "./output"
@@ -60,37 +51,52 @@ client_configs:
       sample_size: True
       num_classes: True
       data_shape: True
-      completeness: True
+      completeness: False
       data_range: True
-      sparsity: True
-      variance: True
-      skewness: True
-      entropy: True
-      kurtosis: True
+      overall_sparsity: True
+      variance: False
+      skewness: False
+      entropy: False
+      kurtosis: False
       class_distribution: True
-      brisque: True
+      brisque: False
       sharpness: False
+      outlier_proportion: True
+      time_to_event_imbalance: False
       plot:
         class_distribution_plot: True
-        data_sample_plot: True
+        data_sample_plot: False
         data_distribution_plot: True
-        class_variance_plot: True
+        class_variance_plot: False
+        feature_correlation_plot: True
+        feature_statistics_plot: True
+      combine:
+        feature_space_distribution: True
 
 server_configs:
-  scheduler: "AsyncScheduler"
+  scheduler: "CompassScheduler"
   scheduler_kwargs:
-    num_clients: 2
+    num_clients: 4
     same_init_model: True
-  aggregator: "FedAsyncAggregator"
+    max_local_steps: 100
+    min_local_steps: 20
+    speed_momentum: 0.9
+    latest_time_factor: 1.5
+  aggregator: "FedCompassAggregator"
   aggregator_kwargs:
-    client_weights_mode: "equal"
-    num_clients: 2
+    weights: "equal"
     staleness_fn: "polynomial"
     staleness_fn_kwargs:
       a: 0.5
     alpha: 0.9
     gradient_based: True
+    num_clients: 4
   device: "cpu"
-  num_global_epochs: 20
+  num_global_epochs: 100
   logging_output_dirname: "./output"
   logging_output_filename: "result"
+  comm_configs:
+    grpc_configs:
+      server_uri: localhost:50051
+      max_message_size: 1048576
+      use_ssl: False
