--- examples/resources/config_gc/flamby/server_fedavg.yaml
@@ -1,91 +1,0 @@
-client_configs:
-  train_configs:
-    # Local trainer
-    trainer: "VanillaTrainer"
-    mode: "step"
-    num_local_steps: 50
-    optim: "Adam"
-    optim_args:
-      lr: 0.1
-    # Loss function
-    loss_fn_path: "./resources/loss/coxloss.py"
-    loss_fn_name: "BaselineLoss"
-    # Client validation
-    do_validation: True
-    do_pre_validation: True
-    metric_path: "./resources/metric/c_index.py"
-    metric_name: "metric"
-    # Differential privacy
-    use_dp: False
-    epsilon: 1
-    clip_grad: False
-    clip_value: 1
-    clip_norm: 1
-    # Data loader
-    train_batch_size: 8
-    val_batch_size: 8
-    train_data_shuffle: True
-    val_data_shuffle: False
-    
-  model_configs:
-    model_path: "./resources/model/fc.py"
-    model_name: "FC"
-    model_kwargs:
-      input_size: 39
-  
-  comm_configs:
-    compressor_configs:
-      enable_compression: False
-      # Used if enable_compression is True
-      lossy_compressor:  "SZ2Compressor"
-      lossless_compressor: "blosc"
-      error_bounding_mode: "REL"
-      error_bound: 1e-3
-      param_cutoff: 1024
-
-  data_readiness_configs:
-    generate_dr_report: True
-    output_dirname: "./output"
-    output_filename: "data_readiness_report"
-    dr_metrics: 
-      class_imbalance: False
-      sample_size: True
-      num_classes: False
-      data_shape: True
-      completeness: True
-      data_range: True
-      sparsity: True
-      variance: True
-      skewness: True
-      entropy: True
-      kurtosis: True
-      class_distribution: False
-      brisque: False
-      sharpness: False
-      outlier_propotion: True
-      plot:
-        class_distribution_plot: False
-        data_sample_plot: False
-        data_distribution_plot: True
-        class_variance_plot: False
-        # time_to_event_plot: True
-      combine:
-        feature_space_distribution: True
-
-server_configs:
-  scheduler: "SyncScheduler"
-  scheduler_kwargs:
-    num_clients: 6
-    same_init_model: True
-  aggregator: "FedAvgAggregator"
-  aggregator_kwargs:
-    client_weights_mode: "equal"
-  device: "cpu"
-  num_global_epochs: 10
-  logging_output_dirname: "./output"
-  logging_output_filename: "result"
-  comm_configs:
-    grpc_configs:
-      server_uri: localhost:50051
-      max_message_size: 1048576
-      use_ssl: False
