+++ examples/resources/configs/flamby/heart_disease/server_fedavg.yaml
@@ -6,14 +6,14 @@ client_configs:
     num_local_steps: 100
     optim: "Adam"
     optim_args:
-      lr: 0.1
+      lr: 0.001
     # Loss function
-    loss_fn_path: "./resources/loss/coxloss.py"
+    loss_fn_path: "./resources/loss/bceloss.py"
     loss_fn_name: "BaselineLoss"
     # Client validation
     do_validation: True
     do_pre_validation: True
-    metric_path: "./resources/metric/c_index.py"
+    metric_path: "./resources/metric/binary_acc.py"
     metric_name: "metric"
     # Differential privacy
     use_dp: False
@@ -22,45 +22,45 @@ client_configs:
     clip_value: 1
     clip_norm: 1
     # Data loader
-    train_batch_size: 64
-    val_batch_size: 64
+    train_batch_size: 4
+    val_batch_size: 4
     train_data_shuffle: True
     val_data_shuffle: False
-  
+
   data_readiness_configs:
     generate_dr_report: True
     output_dirname: "./output"
     output_filename: "data_readiness_report"
     dr_metrics: 
-      class_imbalance: False
+      class_imbalance: True
       sample_size: True
-      num_classes: False
-      data_shape: False
+      num_classes: True
+      data_shape: True
       completeness: False
       data_range: True
-      sparsity: False
+      overall_sparsity: True
       variance: False
       skewness: False
       entropy: False
       kurtosis: False
-      class_distribution: False
+      class_distribution: True
       brisque: False
       sharpness: False
-      outlier_propotion: False
+      outlier_proportion: True
+      time_to_event_imbalance: False
       plot:
-        class_distribution_plot: False
+        class_distribution_plot: True
         data_sample_plot: False
-        data_distribution_plot: False
+        data_distribution_plot: True
         class_variance_plot: False
+        feature_correlation_plot: True
+        feature_statistics_plot: True
       combine:
         feature_space_distribution: True
     
   model_configs:
-    model_path: "./resources/model/fc.py"
-    model_name: "FC"
-    model_kwargs:
-      input_size: 39
-      # num_classes: 1
+    model_path: "./resources/model/flamby_heart_disease.py"
+    model_name: "Baseline"
   
   comm_configs:
     compressor_configs:
@@ -75,13 +75,13 @@ client_configs:
 server_configs:
   scheduler: "SyncScheduler"
   scheduler_kwargs:
-    num_clients: 6
+    num_clients: 2
     same_init_model: True
   aggregator: "FedAvgAggregator"
   aggregator_kwargs:
     client_weights_mode: "equal"
   device: "cpu"
-  num_global_epochs: 10
+  num_global_epochs: 20
   logging_output_dirname: "./output"
   logging_output_filename: "result"
   comm_configs:
