+++ examples/resources/configs/flamby/heart_disease/server_fedcompass.yaml
@@ -4,35 +4,33 @@ client_configs:
     trainer: "VanillaTrainer"
     mode: "step"
     num_local_steps: 100
-    optim: "SGD"
+    optim: "Adam"
     optim_args:
-      lr: 0.01
+      lr: 0.001
     # Loss function
-    loss_fn_path: "./resources/loss/flamby_heart_disease.py"
+    loss_fn_path: "./resources/loss/bceloss.py"
     loss_fn_name: "BaselineLoss"
     # Client validation
     do_validation: True
     do_pre_validation: True
-    metric_path: "./resources/metric/flamby_heart_disease.py"
+    metric_path: "./resources/metric/binary_acc.py"
     metric_name: "metric"
+    send_gradient: True
     # Differential privacy
     use_dp: False
     epsilon: 1
-    clip_grad: True
-    clip_value: 0.5
+    clip_grad: False
+    clip_value: 1
     clip_norm: 1
     # Data loader
-    train_batch_size: 16
-    val_batch_size: 16
+    train_batch_size: 4
+    val_batch_size: 4
     train_data_shuffle: True
     val_data_shuffle: False
 
   model_configs:
     model_path: "./resources/model/flamby_heart_disease.py"
     model_name: "Baseline"
-    # model_kwargs:
-    #   input_size: 13
-    #   num_classes: 1
   
   comm_configs:
     compressor_configs:
@@ -94,7 +92,7 @@ server_configs:
     gradient_based: True
     num_clients: 4
   device: "cpu"
-  num_global_epochs: 17
+  num_global_epochs: 100
   logging_output_dirname: "./output"
   logging_output_filename: "result"
   comm_configs:
