+++ examples/resources/metric/binary_acc.py
@@ -1,0 +1,13 @@
+import numpy as np
+
+def metric(y_true, y_pred):
+    y_true = y_true.astype("uint8")
+    # The try except is needed because when the metric is batched some batches
+    # have one class only
+    try:
+        # return roc_auc_score(y_true, y_pred)
+        # proposed modification in order to get a metric that calcs on center 2
+        # (y=1 only on that center)
+        return ((y_pred > 0.5) == y_true).mean()
+    except ValueError:
+        return np.nan
